Removing any catch page
As per VidWorx's suggestion regarding the sketchy nature of pure redirect sites (ie. that click funnels use this tactic, and it's unnerving for users to see), I'm removing any text from the hiimcurtismc.com page. Instead, this will be purely a redirect w/o any trace of another site in the mix.